feat(S1): Add token formatting and cost calculation helpers
Extract format_commas to shared utils.rs, add calculate_cost ($3/MTok
input, $15/MTok output) and format_token_display helpers used by both
dashboard and stats screens.
