+++ crates/baro-tui/src/screens/execute_completion.rs
@@ -8,6 +8,7 @@ use ratatui::{
 
 use crate::app::App;
 use crate::theme;
+use crate::utils::format_token_display;
 
 pub fn render_completion(f: &mut Frame, app: &App) {
     let area = f.area();
@@ -201,26 +202,10 @@ pub fn render_completion(f: &mut Frame, app: &App) {
         ),
     ]));
 
-    let format_commas = |n: u64| -> String {
-        let s = n.to_string();
-        let mut result = String::new();
-        for (i, c) in s.chars().rev().enumerate() {
-            if i > 0 && i % 3 == 0 {
-                result.push(',');
-            }
-            result.push(c);
-        }
-        result.chars().rev().collect()
-    };
-
     lines.push(Line::from(vec![
-        Span::styled("  Tokens:         ", Style::default().fg(theme::MUTED)),
+        Span::styled("  ", Style::default()),
         Span::styled(
-            format!(
-                "{} in / {} out",
-                format_commas(app.total_input_tokens),
-                format_commas(app.total_output_tokens)
-            ),
+            format_token_display(app.total_input_tokens, app.total_output_tokens),
             Style::default().fg(theme::ACCENT),
         ),
     ]));
